+++ packages/ui/src/components/tables/table-column-customiser/TableColumnCustomiser.cy.ts
@@ -45,9 +45,12 @@ describe.skip("column-customiser-helper.ts", () => {
     value: localStorage
   });
 
-  it("should save the keys of the columns to localStorage if they are just strings", () => {
-    saveCustomisedColumns("test", ["FirstCol", "SecondCol"]);
-    assert.equal(localStorage.getItem("test"), '["FirstCol","SecondCol"]');
+  it("should save the order and selected state of the columns", () => {
+    saveCustomisedColumns("test", [
+      { key: "FirstCol", selected: true },
+      { key: "SecondCol", selected: false }
+    ]);
+    expect(localStorage.getItem("test")).toBe('[{"key":"FirstCol","selected":true},{"key":"SecondCol","selected":false}]');
   });
 
   it("should save the keys of the columns to localStorage when saved with correct ID", () => {
@@ -59,8 +62,22 @@ describe.skip("column-customiser-helper.ts", () => {
     assert.isNull(loadCustomisedColumns("unknownId"));
   });
 
-  it("should load an return a string array if there is data", () => {
-    saveCustomisedColumns("knownId", [{ key: "FirstCol" }, { key: "SecondCol" }]);
-    assert.deepStrictEqual(loadCustomisedColumns("knownId"), ["FirstCol", "SecondCol"]);
+  it("should load an return a ForgeTableColumnSelectedArray if there is data stored as ForgeTableColumnSelected", () => {
+    saveCustomisedColumns("knownId1", [
+      { key: "FirstCol", selected: true },
+      { key: "SecondCol", selected: false }
+    ]);
+    expect(loadCustomisedColumns("knownId1")).toStrictEqual([
+      { key: "FirstCol", selected: true },
+      { key: "SecondCol", selected: false }
+    ]);
+  });
+
+  it("should load an return a ForgeTableColumnSelectedArray if saved data are strings", () => {
+    localStorage.setItem("knownId2", JSON.stringify(["FirstCol", "SecondCol"]));
+    expect(loadCustomisedColumns("knownId2")).toStrictEqual([
+      { key: "FirstCol", selected: true },
+      { key: "SecondCol", selected: true }
+    ]);
   });
 });