+++ packages/ui/src/components/tables/table-column-customiser/column-customiser-helper.ts
@@ -1,15 +1,26 @@
-import { ForgeTableField, ForgeTableFieldArray } from "../../../helpers/types";
+import { ForgeTableFieldKeyed, ForgeTableColumnSelected, ForgeTableColumnSelectedArray } from "../../../helpers/types";
 
-export function getColumnKey(field: string | ({ key: string } & ForgeTableField)): string {
+export function getColumnKey(field: string | ForgeTableFieldKeyed): string {
   return typeof field == "string" ? field : field.key;
 }
 
-export function saveCustomisedColumns(id: string, columns: ForgeTableFieldArray) {
-  window.localStorage.setItem(id, JSON.stringify(columns.map((c) => getColumnKey(c))));
+export function saveCustomisedColumns(id: string, columns: ForgeTableColumnSelectedArray) {
+  window.localStorage.setItem(id, JSON.stringify(columns));
 }
 
-export function loadCustomisedColumns(id: string): string[] | null {
+export function loadCustomisedColumns(id: string): ForgeTableColumnSelectedArray | null {
   const jsonCols = window.localStorage.getItem(id);
+  if (jsonCols != null) {
+    const cols = JSON.parse(jsonCols) as (ForgeTableColumnSelected | string)[];
+    const colsObjects = cols.map((col) => {
+      if (typeof col == "string") {
+        return { key: col, selected: true } as ForgeTableColumnSelected;
+      }
+      return col as ForgeTableColumnSelected;
+    });
 
-  return jsonCols == null ? null : (JSON.parse(jsonCols) as string[]);
+    return colsObjects;
+  } else {
+    return null;
+  }
 }