+++ packages/ui/vite.config.ts
@@ -28,7 +28,7 @@ export default defineConfig(({ mode }) => ({
   plugins: [
     vue(),
     checker({ vueTsc: true }),
-    dts(),
+    dts({ outDir: "dist/types" }),
     istanbul({
       include: "src/*",
       exclude: ["node_modules", "test/"],
@@ -89,7 +89,8 @@ export default defineConfig(({ mode }) => ({
       include: ["src"],
       exclude: ["**/*.cy.ts", "scripts", "**/*.test.ts"],
       reportsDirectory: "./coverage/vitest",
-      reporter: ["cobertura", "html", "json"]
+      reporter: ["cobertura", "html", "json"],
+      provider: "v8"
     }
   }
 }));