+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,37 @@
 # forge-ui-styleguide
 
+## 1.3.1
+
+### Patch Changes
+
+- 9c3c27e: Update ShowClearButton prop to default to false
+- Updated dependencies [9c3c27e]
+  - @3squared/forge-ui@1.7.1
+
+## 1.3.0
+
+### Minor Changes
+
+- 86377bd: Add option to show clear filters button above the table
+
+### Patch Changes
+
+- Updated dependencies [86377bd]
+  - @3squared/forge-ui@1.7.0
+
+## 1.2.0
+
+### Minor Changes
+
+- d59148b: Added a new slot ot the forge table component which allows for extra buttons or other components to be added above the table (and next to pagination)
+
+### Patch Changes
+
+- Updated dependencies [dfef40d]
+- Updated dependencies [7da3033]
+- Updated dependencies [d59148b]
+  - @3squared/forge-ui@1.6.0
+
 ## 1.1.0
 
 ### Minor Changes