+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @3squared/forge-ui
 
+## 1.9.1
+
+### Patch Changes
+
+- bfe129e: Fix typings output
+
+## 1.9.0
+
+### Minor Changes
+
+- 3dea70c: Added props for pagination on the table component for further customization
+
+## 1.8.3
+
+### Patch Changes
+
+- bbd91ed: Update where types are declared in package.json
+
+## 1.8.2
+
+### Patch Changes
+
+- 7a22a46: Add provider to vite config
+
+## 1.8.1
+
+### Patch Changes
+
+- 0241e05: Fix Toggle All padding issue
+
+## 1.8.0
+
+### Minor Changes
+
+- 284f905: Fix ordering issues with the TableColumnCustomiser
+
+## 1.7.1
+
+### Patch Changes
+
+- 9c3c27e: Update ShowClearButton prop to default to false
+
+## 1.7.0
+
+### Minor Changes
+
+- 86377bd: Add option to show clear filters button above the table
+
+## 1.6.0
+
+### Minor Changes
+
+- d59148b: Added a new slot ot the forge table component which allows for extra buttons or other components to be added above the table (and next to pagination)
+
+### Patch Changes
+
+- dfef40d: remove preserve search as default multiselect
+- 7da3033: File Uploader 'Browse Your Computer' Hover not working because the button wasn't hovered
+
 ## 1.5.0
 
 ### Minor Changes